refactor: move to helper
