+++ packages/evm-helpers/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/contract';
 export * from './lib/types';
+export * from './lib/metadata';
